requirements.txt
pyrogram==2.0.106
TgCrypto
py-tgcalls==0.9.9
yt-dlp
python-dotenv
